# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/.well-known/openid-configuration": {
      "get": {
        "tags": [
          "oidc"
        ],
        "summary": "Discovery",
        "operationId": "discovery__well_known_openid_configuration_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/.well-known/jwks.json": {
      "get": {
        "tags": [
          "oidc"
        ],
        "summary": "Jwks",
        "operationId": "jwks__well_known_jwks_json_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/authorize": {
      "get": {
        "tags": [
          "oidc"
        ],
        "summary": "Authorize",
        "operationId": "authorize_authorize_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/interaction/{uid}": {
      "get": {
        "tags": [
          "interaction"
        ],
        "summary": "Get Interaction",
        "operationId": "get_interaction_interaction__uid__get",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Uid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionView"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or expired uid"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/interaction/{uid}/login": {
      "post": {
        "tags": [
          "interaction"
        ],
        "summary": "Login",
        "description": "On success, sets the `__Host-session` cookie and answers the next step.",
        "operationId": "login_interaction__uid__login_post",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Uid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionView"
                }
              }
            }
          },
          "401": {
            "description": "Bad numero_fiscal or password (same answer for both)"
          },
          "404": {
            "description": "Unknown or expired uid"
          },
          "423": {
            "description": "Account locked"
          },
          "429": {
            "description": "Rate limited"
          },
          "501": {
            "description": "Not built yet (B1)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/interaction/{uid}/consent": {
      "post": {
        "tags": [
          "interaction"
        ],
        "summary": "Consent",
        "description": "approve=false ends the flow: redirect_to then carries error=access_denied.",
        "operationId": "consent_interaction__uid__consent_post",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Uid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionView"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or expired uid"
          },
          "501": {
            "description": "Not built yet (B1)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/keys/rotate": {
      "post": {
        "tags": [
          "key"
        ],
        "summary": "Rotate",
        "operationId": "rotate_admin_keys_rotate_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/health/live": {
      "get": {
        "tags": [
          "ops"
        ],
        "summary": "Live",
        "operationId": "live_health_live_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "tags": [
          "ops"
        ],
        "summary": "Ready",
        "operationId": "ready_health_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConsentBody": {
        "properties": {
          "approve": {
            "type": "boolean",
            "title": "Approve"
          }
        },
        "type": "object",
        "required": [
          "approve"
        ],
        "title": "ConsentBody"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InteractionView": {
        "properties": {
          "prompt": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "login",
                  "consent"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt",
            "description": "Step to display next. null: finished, follow redirect_to"
          },
          "client_name": {
            "type": "string",
            "title": "Client Name"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes"
          },
          "redirect_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Redirect To",
            "description": "Set when prompt is null: back to /authorize, which issues the code"
          }
        },
        "type": "object",
        "required": [
          "prompt",
          "client_name",
          "scopes"
        ],
        "title": "InteractionView"
      },
      "LoginBody": {
        "properties": {
          "numero_fiscal": {
            "type": "string",
            "pattern": "^\\d{13}$",
            "title": "Numero Fiscal"
          },
          "password": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "numero_fiscal",
          "password"
        ],
        "title": "LoginBody"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}